refactor: extract shared tree-sitter logic to TreeSitterTokenProcessor
Create TreeSitterTokenProcessor utility with shared methods:
- extractTokens(from:code:query:)
- byteRangeToStringRange(_:in:)
- mapCaptureToTokenType(_:)
- deduplicateAndSort(_:)
- renderTokensToHTML(code:tokens:)

This eliminates ~100 lines of duplicate code between TreeSitterHighlighter
and LazyTreeSitterHighlighter.

Closes #24
